requirements.txt
ckanapi
requests>=2.4.2
pytest
pytest-ckan